Add dependency resolution to the source generator itself
The wrapping `DependencyResolverGenerator` means generated filenames also
show up as coming from that generator, rather than `StuntSourceGenerator`.

The added value was very little anyway, so just move it all to the project.
